Work around the strange behavior of the sssd 'simple' access_provider
When a realm is in the allow-permitted-logins mode, but no such
logins have been set, the simple_allow_users is empty. sssd treats
this as if the line is not present, and allows any login.

Setting the value to a comma gets around this problem. Further discussion
will take place on the sssd mailing list.

https://bugs.freedesktop.org/show_bug.cgi?id=56027