Limit Netbios name to 15 chars when joining AD domain
In the future we will have more code for managing the computer name.

'net ads join' requires us to setup 'netbios name' to a short name
in smb.conf in order to do a join with a long host name.

We also need to setup sssd.conf to authenticate correctly when the
hostname doesn't match the computer account name. We use
ldap_sasl_authid to accomplish this.

https://bugs.freedesktop.org/show_bug.cgi?id=69016