Disable automatic AD joins by default
We only offer them as an option to clients if they have been configured
in the realmd.conf file.

This is because automatic AD joins do not have the mutual authentication
we usually expect with kerberos. The computer account secret is
predictable and not secure enough to be on by default.

https://bugs.freedesktop.org/show_bug.cgi?id=89205