Be more liberal on what we accept as a domain name
Make the checks on what we accept as a domain name more liberal
for values coming in from the network.

DNS Domain names are pretty liberal (internet domain names
are more restrictive) See RFC 2181 section 11

http://www.ietf.org/rfc/rfc2181.txt

However we cannot consume names with whitespace and problematic
punctuation, due to the various programs that parse the
configuration files we set up.